requirements.txt
# Core
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0

# LLM
anthropic>=0.40.0

# Voice
elevenlabs>=1.0.0
deepgram-sdk>=3.0.0
# openai-whisper>=20231117  # Optional: local Whisper

# WebSocket
websockets>=12.0

# Database
supabase>=2.0.0

# Utilities
httpx>=0.26.0
python-multipart>=0.0.6
pyyaml>=6.0
